Now “Task Collector” has only “Done” in task mark cycle.
Removed “Selected” (`*`) and “In Progress” (`/`).

Having keyboard shortcuts [^1] to cycle through all the task statuses
proved to be not very useful as most of the time a separate keyboard
shortcut is useful when marking multiple tasks in rapid succession.
Those will mostly be small tasks, such as daily bootstrap tasks, which
are never marked as “Selected” or “In Progress”.

[^1]: ⌘⌥M and ⌘⌥⇧M

Marking as selected or in-progress is only done at the moment of making
the decision for larger tasks, so ⌘M followed by `*` or `/` will not be
a problem.